Refactoring ogip
